+++ app/controllers/players_controller.rb
@@ -989,7 +989,7 @@ class PlayersController < ApplicationController
 
           current_teams = club.current_teams
           current_licenses = (player.current_licenses || []).reject do |l|
-                               [6, 7].include?(l['history'].last['license_status_id'].to_i)
+                               [License::TRANSFER, License::IGNORED].include?(LicenseEffectiveStatus.current_status_id(l))
                              end.map { |l| l['team_id'] }
 
           # check for licenses for that club